+++ src/components/mod.rs
@@ -35,13 +35,13 @@ where
 	Some(parsed)
 }
 
-pub(crate) fn collect_day_and_validate(s: &str, position: &mut usize, month: u8) -> Option<u8> {
+pub(crate) fn collect_day_and_validate(s: &str, position: &mut usize, month: u32) -> Option<u32> {
 	let parsed_day = collect_ascii_digits(s, position);
 	if parsed_day.len() != 2 {
 		return None;
 	}
 
-	let day = parsed_day.parse::<u8>().ok()?;
+	let day = parsed_day.parse::<u32>().ok()?;
 	let max_days = max_days_in_month_year(month, 4).unwrap();
 	if !(1..=max_days).contains(&day) {
 		return None;
@@ -50,13 +50,13 @@ pub(crate) fn collect_day_and_validate(s: &str, position: &mut usize, month: u8)
 	Some(day)
 }
 
-pub(crate) fn collect_month_and_validate(s: &str, position: &mut usize) -> Option<u8> {
+pub(crate) fn collect_month_and_validate(s: &str, position: &mut usize) -> Option<u32> {
 	let parsed_month = collect_ascii_digits(s, position);
 	if parsed_month.len() != 2 {
 		return None;
 	}
 
-	let month = parsed_month.parse::<u8>().ok()?;
+	let month = parsed_month.parse::<u32>().ok()?;
 	if !is_valid_month(&month) {
 		return None;
 	}
